parser_service/requirements.txt
alembic==1.9.0
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
envparse==0.2.0
fastapi==0.88.0
flake8==6.0.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
